requirements.txt
flask
google-generativeai
boto3